whisper-nikud/pyproject.toml
[project]
name = "whisper-heb-nikud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "datasets==2.21.0",
    "evaluate>=0.4.5",
    "jiwer>=4.0.0",
    "librosa>=0.11.0",
    "openai-whisper>=20250625",
    "soundfile>=0.13.1",
    "torchaudio>=2.8.0",
    "transformers>=4.52.0",
]

[dependency-groups]
dev = [
    "pandas>=3.0.0",
    "pydub>=0.25.1",
    "tensorboard>=2.20.0",
    "wandb>=0.21.4",
    "huggingface-hub>=0.35.0",
    "gradio>=5.46.0",
    "accelerate>=1.12.0",
]
